calculator.py
a program taking two integer input as x and y and printing all the basic mathe matical operations. there may be some error in answers related to division because of restriction of answers to integer form.
i prefer you to take the first value bigger than the second to prevent any error.